# modules/gateway/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bedrock Gateway API
  description: |
    Multi-tenant SaaS proxy for Amazon Bedrock with Cognito authentication,
    cascading budget management, rate limiting, and an admin dashboard.

    ## Authentication

    All API endpoints require authentication via one of:
    - **Cognito JWT Bearer Token**: `Authorization: Bearer <access_token>`
    - **Legacy Token Exchange**: `POST /auth/exchange` (deprecated)

    Human users authenticate via Cognito PKCE flow using the `bg-cognito-auth.sh` CLI.
    Agents/services use Cognito `client_credentials` flow for M2M authentication.

    ## Rate Limiting

    Rate limits are enforced at organization, department, team, and user levels.
    When rate limited, responses include:
    - `Retry-After` header
    - `X-RateLimit-Limit`, `X-RateLimit-Remaining`, `X-RateLimit-Reset` headers

    ## Budget Enforcement

    Budget limits cascade from organization to user level.
    When budget is exceeded (hard mode), requests are rejected with 429.
    Response headers include: `X-Budget-Remaining-USD`, `X-Budget-Period`
  version: 1.0.0
  contact:
    name: Platform Team
    email: platform@company.com
  license:
    name: Private
    url: https://company.com/license
servers:
  - url: https://{cloudfront_domain}/api
    description: Production Gateway via CloudFront
    variables:
      cloudfront_domain:
        default: gateway.company.com
        description: CloudFront distribution domain
...
security:
  - bearerAuth: []
...
paths:
...
  /budgets:
    post:
      tags: [budgets]
      summary: Create budget configuration
      operationId: createBudget
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BudgetCreateRequest'
      responses:
        '201':
          description: Budget created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetResponse'
  /budgets/{budget_id}:
    get:
      tags: [budgets]
      summary: Get budget configuration
      operationId: getBudget
      parameters:
        - name: budget_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Budget configuration
    put:
      tags: [budgets]
      summary: Update budget configuration
      operationId: updateBudget
      parameters:
        - name: budget_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BudgetUpdateRequest'
      responses:
        '200':
          description: Budget updated
    delete:
      tags: [budgets]
      summary: Delete budget configuration
      operationId: deleteBudget
      parameters:
        - name: budget_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Budget deleted
  /budgets/entity/{entity_type}/{entity_id}:
    get:
      tags: [budgets]
      summary: Get budgets for entity
      operationId: getBudgetsForEntity
      parameters:
        - name: entity_type
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/EntityType'
        - name: entity_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Budget configurations
  /budgets/status/{entity_type}/{entity_id}:
    get:
      tags: [budgets]
      summary: Get budget status
      operationId: getBudgetStatus
      parameters:
        - name: entity_type
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/EntityType'
        - name: entity_id
          in: path
          required: true
          schema:
            type: string
        - name: period_type
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/PeriodType'
      responses:
        '200':
          description: Budget status
  /budgets/usage/{entity_type}/{entity_id}:
    get:
      tags: [budgets]
      summary: Get budget usage
      operationId: getBudgetUsage
      parameters:
        - name: entity_type
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/EntityType'
        - name: entity_id
          in: path
          required: true
          schema:
            type: string
        - name: period_type
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/PeriodType'
      responses:
        '200':
          description: Budget usage
  /budgets/calculate-cost:
    post:
      tags: [budgets]
      summary: Calculate cost for token usage
      operationId: calculateCost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CostCalculationRequest'
      responses:
        '200':
          description: Cost calculation
  /budgets/record-cost:
    post:
      tags: [budgets]
      summary: Record cost against budgets
      operationId: recordCost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CostRecordRequest'
      responses:
        '204':
          description: Cost recorded
...
  /budgets/organization/alerts:
    get:
      tags: [budgets]
      summary: Get budget alerts
      operationId: getBudgetAlerts
      parameters:
        - name: threshold_percent
          in: query
          schema:
            type: number
            default: 80
      responses:
        '200':
          description: Budget alerts

  # ==========================================================================
  # Rate Limit Endpoints
  # ==========================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Cognito JWT access token
  schemas:
...
    EntityType:
      type: string
      enum:
        - org
        - department
        - team
        - user
        - service_account
    PeriodType:
      type: string
      enum:
        - daily
        - weekly
        - monthly
    EnforcementMode:
      type: string
      enum:
        - soft
        - hard
    BudgetCreateRequest:
      type: object
      required:
        - entity_type
        - entity_id
        - period_type
        - budget_amount_usd
      properties:
        entity_type:
          $ref: '#/components/schemas/EntityType'
        entity_id:
          type: string
        period_type:
          $ref: '#/components/schemas/PeriodType'
        budget_amount_usd:
          type: number
          minimum: 0
        enforcement_mode:
          $ref: '#/components/schemas/EnforcementMode'
          default: hard
    BudgetUpdateRequest:
      type: object
      properties:
        budget_amount_usd:
          type: number
          minimum: 0
        enforcement_mode:
          $ref: '#/components/schemas/EnforcementMode'
    BudgetResponse:
      type: object
      properties:
        id:
          type: string
        entity_type:
          $ref: '#/components/schemas/EntityType'
        entity_id:
          type: string
        period_type:
          $ref: '#/components/schemas/PeriodType'
        budget_amount_usd:
          type: number
        enforcement_mode:
          $ref: '#/components/schemas/EnforcementMode'
        org_id:
          type: string
        updated_at:
          type: string
          format: date-time
    CostCalculationRequest:
      type: object
      required:
        - model_name
        - tokens_in
        - tokens_out
      properties:
        model_name:
          type: string
        tokens_in:
          type: integer
          minimum: 0
        tokens_out:
          type: integer
          minimum: 0
    CostRecordRequest:
      type: object
      required:
        - entity_type
        - entity_id
        - model_name
        - tokens_in
        - tokens_out
      properties:
        entity_type:
          $ref: '#/components/schemas/EntityType'
        entity_id:
          type: string
        model_name:
          type: string
        tokens_in:
          type: integer
        tokens_out:
          type: integer
        request_cost_usd:
          type: number

    # Rate Limit Schemas
...